+++ desktop-app/scripts/experimental/fresh-install.sh
@@ -1,0 +1,2 @@
+rm -rf /Users/liamz/Library/Application\ Support/Dappnet/
+sudo rm -rf /Applications/Dappnet.app